-added support for specialization constants -removed deprecated shader files
Former-commit-id: fa9c21bf4ecdd4aba33aee657a334a6eb991b775

+++ VEngine/Code/Graphics/RendererConsts.h
@@ -4,19 +4,19 @@ namespace VEngine
 {
 	namespace RendererConsts
 	{
-		constexpr size_t FRAMES_IN_FLIGHT = 2;
-		constexpr size_t TEXTURE_ARRAY_SIZE = 1024;
-		constexpr size_t Z_BINS = 8192;
-		constexpr size_t LIGHTING_TILE_SIZE = 16;
-		constexpr size_t LUMINANCE_HISTOGRAM_SIZE = 256;
-		constexpr size_t MAX_TAA_HALTON_SAMPLES = 16;
-		constexpr size_t MAX_MATERIALS = 32 * 1024;
-		constexpr size_t MAX_VERTICES = 4 * 1024 * 1024; // ~4 million vertices
-		constexpr size_t MAX_INDICES = MAX_VERTICES * 3;
-		constexpr size_t STAGING_BUFFER_SIZE = 128 * 1024 * 1024; 
-		constexpr size_t MAX_SUB_MESHES = 32 * 1024;
-		constexpr size_t MAPPABLE_UBO_BLOCK_SIZE = 1024 * 1024;
-		constexpr size_t MAPPABLE_SSBO_BLOCK_SIZE = 8 * 1024 * 1024;
+		constexpr uint32_t FRAMES_IN_FLIGHT = 2;
+		constexpr uint32_t TEXTURE_ARRAY_SIZE = 1024;
+		constexpr uint32_t Z_BINS = 8192;
+		constexpr uint32_t LIGHTING_TILE_SIZE = 16;
+		constexpr uint32_t LUMINANCE_HISTOGRAM_SIZE = 256;
+		constexpr uint32_t MAX_TAA_HALTON_SAMPLES = 16;
+		constexpr uint32_t MAX_MATERIALS = 32 * 1024;
+		constexpr uint32_t MAX_VERTICES = 4 * 1024 * 1024; // ~4 million vertices
+		constexpr uint32_t MAX_INDICES = MAX_VERTICES * 3;
+		constexpr uint32_t STAGING_BUFFER_SIZE = 128 * 1024 * 1024; 
+		constexpr uint32_t MAX_SUB_MESHES = 32 * 1024;
+		constexpr uint32_t MAPPABLE_UBO_BLOCK_SIZE = 1024 * 1024;
+		constexpr uint32_t MAPPABLE_SSBO_BLOCK_SIZE = 8 * 1024 * 1024;
 		constexpr uint32_t TRIANGLE_FILTERING_CLUSTER_SIZE = 256;
 		constexpr float Z_BIN_DEPTH = 1.0f;
 	}
